Introduce tablename objects (#13)
* Fix units and trivial implementation issues.

This is temporarily introducing an inelegant design where the name
of a query template must be specified twice. This is an unfortunate
byproduct of the way I manage to gradually introduce the proposed API
and design changes. This is expected to be resolved over the next few
PRs.

* Fix CI with coverage.

* Added class TableName.

This allow to formally type strings representing table name and
opens the door for wrapper classes to SQLAlchemy (or other ORMs).

* f-string formatting not available in Python <= 3.5.

* Use method name compile instead of render for consistency, added
docstrings.